Survey project administration improvements

+++ src/form/sections/sections.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   ConflictException,
   HttpException,
   HttpStatus,
@@ -10,7 +11,7 @@ import {
 import { PrismaService } from "src/prisma/prisma.service";
 import { ConditionDto } from "./dto/change-next-section.dto";
 import { QuestionsService } from "../questions/questions.service";
-import { Prisma, QuestionSection } from "@prisma/client";
+import { Prisma, QuestionSection, QuestionType } from "@prisma/client";
 import { NotFoundError } from "@prisma/client/runtime";
 import { PrismaError } from "prisma-error-enum";
 
@@ -63,24 +64,33 @@ export class SectionsService {
    * Gets all registered sections in the database without their questions
    */
   async allSections() {
-    return await this.prisma.questionSection.findMany({});
+    return await this.prisma.questionSection.findMany({
+      orderBy: {
+        id: "asc"
+      }
+    });
   }
 
   /**
    * Gets all registered sections in the database with their questions in order
    */
   async allSectionsWithOrderedQuestions() {
-    const sections = await this.prisma.questionSection.findMany({});
+    const sections = await this.prisma.questionSection.findMany({
+      orderBy: {
+        id: "asc"
+      }
+    });
     const result = await Promise.all(
       sections.map(async (section) => {
-        delete section["nextSectionId"];
-        return {
+        const finalSection = {
           ...section,
           questions: await this.questions.getQuestionsBySectionInOrder(
             section.id
           ),
           next: await this.getSectionNext(section)
         };
+        delete finalSection["nextSectionId"];
+        return finalSection;
       })
     );
     return result;
@@ -101,7 +111,7 @@ export class SectionsService {
   async firstSection() {
     try {
       const firstSection = await this.prisma.questionSection.findFirstOrThrow({
-        where: { previousSection: null, conditioned: { none: {} } }
+        where: { previousSections: { none: {} }, conditioned: { none: {} } }
       });
       return this.section(firstSection.id);
     } catch (error) {
@@ -161,16 +171,29 @@ export class SectionsService {
   /**
    * Sets the next section directly to another section
    * @param id ID of the target section
-   * @param nextSection ID of the next section
+   * @param nextSectionId ID of the next section
    */
-  async setSectionNext(id: number, nextSection: number) {
+  async setSectionNext(id: number, nextSectionId: number | null) {
+    const section = await this.prisma.questionSection.findUniqueOrThrow({
+      where: { id }
+    });
+
+    if (nextSectionId) {
+      const nextSection = await this.prisma.questionSection.findUniqueOrThrow({
+        where: { id: nextSectionId }
+      });
+
+      if (nextSection.id === section.id) {
+        throw new ConflictException("You cannot point to the section itself");
+      }
+    }
+
     try {
       await this.prisma.$transaction(async (tx) => {
         await tx.condition.deleteMany({
           where: {
             question: {
-              section: { id },
-              conditions: { some: {} }
+              section: { id }
             }
           }
         });
@@ -179,9 +202,13 @@ export class SectionsService {
           where: { id },
           data: {
             nextSection: {
-              connect: {
-                id: nextSection
-              }
+              ...(nextSectionId === null
+                ? { disconnect: true }
+                : {
+                    connect: {
+                      id: nextSectionId
+                    }
+                  })
             }
           }
         });
@@ -212,11 +239,25 @@ export class SectionsService {
         where: { id: condition.question, section: { id } },
         include: { answers: true, section: true }
       });
+
+      // Check if question is single choice
+      if (question.type !== QuestionType.SINGLE_CHOICE) {
+        throw new BadRequestException(
+          "The question provided is not a single choice question"
+        );
+      }
+
+      // Check if answers are correct
       const answersIds = question.answers.map((a) => a.id);
       if (
-        !Object.keys(condition.answers).every((answer) =>
-          answersIds.includes(+answer)
-        )
+        !Object.keys(condition.answers).every((answer) => {
+          // Check if any answer does point to the same question section
+          if (condition.answers[answer] === question.section.id)
+            throw new ConflictException(
+              "You cannot point on the section itself"
+            );
+          return answersIds.includes(+answer);
+        })
       ) {
         throw new NotFoundException(
           "The answers provided could not be found in the question provided"
@@ -225,6 +266,13 @@ export class SectionsService {
 
       // Create the conditions
       await this.prisma.$transaction(async (tx) => {
+        await tx.question.update({
+          where: { id: condition.question },
+          data: {
+            required: true
+          }
+        });
+
         await tx.condition.createMany({
           data: Object.entries(condition.answers).map(([answer, section]) => {
             return {
@@ -240,7 +288,7 @@ export class SectionsService {
             where: { id },
             data: {
               nextSection: {
-                delete: true
+                disconnect: true
               }
             }
           });
